requirements.txt
numpy
pandas
scikit-learn
streamlit
pillow
tensorflow
tqdm
joblib

